server/requirements.txt
flask==3.0.0
supabase==2.3.0
python-dotenv==1.0.0
pyjwt==2.8.0
requests==2.31.0
google-auth==2.25.2

